Make all pre-releases dry-runs (drafts deleted afterwards) by default
Doing GitHub Release for every single build is very wasteful on GH resources and I am ending this practice. Technically we could tag every release when built and just check a tag exists, but this is also unnecessary - instead we just ensure we run all the tests every time we do a manually triggered release.